[PLAYER-5446] Move main classes and panels common styles
