+++ sdk/react/src/panels/styles/common.styles.json
